+++ UBB--SE-2025-Workout-main/NeoIsisJob/Workout.Core/IRepositories/IUserRepository.cs
@@ -5,24 +5,24 @@ namespace ServerLibraryProject.Interfaces
 {
     public interface IUserRepository
     {
-        void Follow(long userId, long whoToFollowId);
+        void Follow(int userId, int whoToFollowId);
 
         List<UserModel> GetAll();
 
-        UserModel GetById(long id);
+        UserModel GetById(int id);
 
         UserModel? GetByUsername(string username);
 
-        List<UserModel> GetUserFollowers(long id);
+        List<UserModel> GetUserFollowers(int id);
 
-        List<UserModel> GetUserFollowing(long id);
+        List<UserModel> GetUserFollowing(int id);
 
         UserModel Save(UserModel entity);
 
-        void Unfollow(long userId, long whoToUnfollowId);
+        void Unfollow(int userId, int whoToUnfollowId);
 
-        void JoinGroup(long userId, long groupId);
+        void JoinGroup(int userId, long groupId);
 
-        void ExitGroup(long userId, long groupId);
+        void ExitGroup(int userId, long groupId);
     }
 }
